Remove `notes/new/` page, move creation to `edit`.
The `notes/new/` and `notes/edit` pages were almost entirely duplicates
of each other.
This change refactors `notes/edit` to allow using the noteid `new`
to create a note.
The `createNote` function has been moved appropriately and the
`deleteNote` function updated to only delete existing notes.

This commit also removes `notes/new` as obsolete.
